+++ src/ugly_widget/mod.rs
@@ -1,3 +1,2 @@
-pub mod combo_box;
 pub mod radio_button;
 pub mod ugly_list;
